google-adk-tutorial/app/requirements.txt
google-adk>=0.8.0
litellm>=1.10.0
python-dotenv>=1.0.0 